refactor: move ExponentialBackoffManager to internal
