+++ server/src/main/java/org/apache/kafka/server/util/NetworkPartitionMetadataClient.java
@@ -30,7 +30,7 @@ import org.apache.kafka.common.network.ListenerName;
 import org.apache.kafka.common.protocol.Errors;
 import org.apache.kafka.common.requests.ListOffsetsRequest;
 import org.apache.kafka.common.requests.ListOffsetsResponse;
-import org.apache.kafka.common.utils.ExponentialBackoffManager;
+import org.apache.kafka.common.utils.internals.ExponentialBackoffManager;
 import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.common.utils.Utils;
 import org.apache.kafka.metadata.MetadataCache;
