+++ server/src/test/java/org/apache/kafka/server/util/NetworkPartitionMetadataClientTest.java
@@ -32,7 +32,7 @@ import org.apache.kafka.common.protocol.Errors;
 import org.apache.kafka.common.requests.ListOffsetsRequest;
 import org.apache.kafka.common.requests.ListOffsetsResponse;
 import org.apache.kafka.common.security.auth.SecurityProtocol;
-import org.apache.kafka.common.utils.ExponentialBackoffManager;
+import org.apache.kafka.common.utils.internals.ExponentialBackoffManager;
 import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.metadata.MetadataCache;
 import org.apache.kafka.server.util.timer.MockTimer;
